+++ goldie.go
@@ -9,9 +9,6 @@
 package goldie
 
 import (
-	"bytes"
-	"encoding/json"
-	"encoding/xml"
 	"flag"
 	"fmt"
 	"io/ioutil"
@@ -19,10 +16,7 @@ import (
 	"path/filepath"
 	"strings"
 	"testing"
-	"text/template"
 	"time"
-
-	"errors"
 
 	"github.com/pmezard/go-difflib/difflib"
 	"github.com/sergi/go-diff/diffmatchpatch"
@@ -78,7 +72,9 @@ var (
 	ts = time.Now()
 )
 
-type goldie struct {
+// Goldie is the root structure for the test runner. It provides test assertions based on golden files. It's
+// typically used for testing responses with larger data bodies.
+type Goldie struct {
 	fixtureDir     string
 	fileNameSuffix string
 	filePerms      os.FileMode
@@ -95,8 +91,8 @@ type goldie struct {
 
 // New creates a new golden file tester. If there is an issue with applying any
 // of the options, an error will be reported and t.FailNow() will be called.
-func New(t *testing.T, options ...Option) *goldie {
-	g := goldie{
+func New(t *testing.T, options ...Option) *Goldie {
+	g := Goldie{
 		fixtureDir:           defaultFixtureDir,
 		fileNameSuffix:       defaultFileNameSuffix,
 		filePerms:            defaultFilePerms,
@@ -111,7 +107,7 @@ func New(t *testing.T, options ...Option) *goldie {
 	for _, option := range options {
 		err = option(&g)
 		if err != nil {
-			t.Error(fmt.Errorf("Could not apply option: %w", err))
+			t.Error(fmt.Errorf("could not apply option: %w", err))
 			t.FailNow()
 		}
 	}
@@ -149,222 +145,12 @@ func Diff(engine DiffEngine, actual string, expected string) string {
 	return diff
 }
 
-// === OptionProcessor ===============================
-
-// WithFixtureDir sets the fixture directory.
-//
-// Defaults to `testdata`
-func (g *goldie) WithFixtureDir(dir string) error {
-	g.fixtureDir = dir
-	return nil
-}
-
-// WithNameSuffix sets the file suffix to be used for the golden file.
-//
-// Defaults to `.golden`
-func (g *goldie) WithNameSuffix(suffix string) error {
-	g.fileNameSuffix = suffix
-	return nil
-}
-
-// WithFilePerms sets the file permissions on the golden files that are
-// created.
-//
-// Defaults to 0644.
-func (g *goldie) WithFilePerms(mode os.FileMode) error {
-	g.filePerms = mode
-	return nil
-}
-
-// WithDirPerms sets the directory permissions for the directories in which the
-// golden files are created.
-//
-// Defaults to 0755.
-func (g *goldie) WithDirPerms(mode os.FileMode) error {
-	g.dirPerms = mode
-	return nil
-}
-
-// WithDiffEngine sets the `diff` engine that will be used to generate the
-// `diff` text.
-func (g *goldie) WithDiffEngine(engine DiffEngine) error {
-	g.diffEngine = engine
-	return nil
-}
-
-// WithDiffFn sets the `diff` engine to be a function that implements the
-// DiffFn signature. This allows for any customized diff logic you would like
-// to create.
-func (g *goldie) WithDiffFn(fn DiffFn) error {
-	g.diffFn = fn
-	return nil
-}
-
-// WithIgnoreTemplateErrors allows template processing to ignore any variables
-// in the template that do not have corresponding data values passed in.
-//
-// Default value is false.
-func (g *goldie) WithIgnoreTemplateErrors(ignoreErrors bool) error {
-	g.ignoreTemplateErrors = ignoreErrors
-	return nil
-}
-
-// WithTestNameForDir will create a directory with the test's name in the
-// fixture directory to store all the golden files.
-//
-// Default value is false.
-func (g *goldie) WithTestNameForDir(use bool) error {
-	g.useTestNameForDir = use
-	return nil
-}
-
-// WithSubTestNameForDir will create a directory with the sub test's name to
-// store all the golden files. If WithTestNameForDir is enabled, it will be in
-// the test name's directory. Otherwise, it will be in the fixture directory.
-//
-// Default value is false.
-func (g *goldie) WithSubTestNameForDir(use bool) error {
-	g.useSubTestNameForDir = use
-	return nil
-}
-
-// Assert compares the actual data received with the expected data in the
-// golden files. If the update flag is set, it will also update the golden
-// file.
-//
-// `name` refers to the name of the test and it should typically be unique
-// within the package. Also it should be a valid file name (so keeping to
-// `a-z0-9\-\_` is a good idea).
-func (g *goldie) Assert(t *testing.T, name string, actualData []byte) {
-	t.Helper()
-	if *update {
-		err := g.Update(t, name, actualData)
-		if err != nil {
-			t.Error(err)
-			t.FailNow()
-		}
-	}
-
-	err := g.compare(t, name, actualData)
-	if err != nil {
-		{
-			var e *errFixtureNotFound
-			if errors.As(err, &e) {
-				t.Error(err)
-				t.FailNow()
-				return
-			}
-		}
-
-		{
-			var e *errFixtureMismatch
-			if errors.As(err, &e) {
-				t.Error(err)
-				return
-			}
-		}
-
-		t.Error(err)
-	}
-}
-
-// AssertJson compares the actual json data received with expected data in the
-// golden files. If the update flag is set, it will also update the golden
-// file.
-//
-// `name` refers to the name of the test and it should typically be unique
-// within the package. Also it should be a valid file name (so keeping to
-// `a-z0-9\-\_` is a good idea).
-func (g *goldie) AssertJson(t *testing.T, name string, actualJsonData interface{}) {
-	t.Helper()
-	js, err := json.MarshalIndent(actualJsonData, "", "  ")
-
-	if err != nil {
-		t.Error(err)
-		t.FailNow()
-	}
-
-	g.Assert(t, name, normalizeLF(js))
-}
-
-// AssertXml compares the actual xml data received with expected data in the
-// golden files. If the update flag is set, it will also update the golden
-// file.
-//
-// `name` refers to the name of the test and it should typically be unique
-// within the package. Also it should be a valid file name (so keeping to
-// `a-z0-9\-\_` is a good idea).
-func (g *goldie) AssertXml(t *testing.T, name string, actualXmlData interface{}) {
-	t.Helper()
-	x, err := xml.MarshalIndent(actualXmlData, "", "  ")
-
-	if err != nil {
-		t.Error(err)
-		t.FailNow()
-	}
-
-	g.Assert(t, name, normalizeLF(x))
-}
-
-// normalizeLF normalizes line feed character set across os (es)
-// \r\n (windows) & \r (mac) into \n (unix)
-func normalizeLF(d []byte) []byte {
-	// if empty / nil return as is
-	if len(d) == 0 {
-		return d
-	}
-	// replace CR LF \r\n (windows) with LF \n (unix)
-	d = bytes.Replace(d, []byte{13, 10}, []byte{10}, -1)
-	// replace CF \r (mac) with LF \n (unix)
-	d = bytes.Replace(d, []byte{13}, []byte{10}, -1)
-	return d
-}
-
-// Assert compares the actual data received with the expected data in the
-// golden files after executing it as a template with data parameter. If the
-// update flag is set, it will also update the golden file.  `name` refers to
-// the name of the test and it should typically be unique within the package.
-// Also it should be a valid file name (so keeping to `a-z0-9\-\_` is a good
-// idea).
-func (g *goldie) AssertWithTemplate(t *testing.T, name string, data interface{}, actualData []byte) {
-	t.Helper()
-	if *update {
-		err := g.Update(t, name, actualData)
-		if err != nil {
-			t.Error(err)
-			t.FailNow()
-		}
-	}
-
-	err := g.compareTemplate(t, name, data, actualData)
-	if err != nil {
-		{
-			var e *errFixtureNotFound
-			if errors.As(err, &e) {
-				t.Error(err)
-				t.FailNow()
-				return
-			}
-		}
-
-		{
-			var e *errFixtureMismatch
-			if errors.As(err, &e) {
-				t.Error(err)
-				return
-			}
-		}
-
-		t.Error(err)
-	}
-}
-
 // Update will update the golden fixtures with the received actual data.
 //
 // This method does not need to be called from code, but it's exposed so that
 // it can be explicitly called if needed. The more common approach would be to
 // update using `go test -update ./...`.
-func (g *goldie) Update(t *testing.T, name string, actualData []byte) error {
+func (g *Goldie) Update(t *testing.T, name string, actualData []byte) error {
 	goldenFile := g.GoldenFileName(t, name)
 	goldenFileDir := filepath.Dir(goldenFile)
 	if err := g.ensureDir(goldenFileDir); err != nil {
@@ -382,83 +168,8 @@ func (g *goldie) Update(t *testing.T, name string, actualData []byte) error {
 	return nil
 }
 
-// compare is reading the golden fixture file and compare the stored data with
-// the actual data.
-func (g *goldie) compare(t *testing.T, name string, actualData []byte) error {
-	expectedData, err := ioutil.ReadFile(g.GoldenFileName(t, name))
-
-	if err != nil {
-		if os.IsNotExist(err) {
-			return newErrFixtureNotFound()
-		}
-
-		return fmt.Errorf("Expected %s to be nil", err.Error())
-	}
-
-	if !bytes.Equal(actualData, expectedData) {
-		msg := "Result did not match the golden fixture. Diff is below:\n\n"
-		actual := string(actualData)
-		expected := string(expectedData)
-
-		if g.diffFn != nil {
-			msg += g.diffFn(actual, expected)
-		} else {
-			msg += Diff(g.diffEngine, actual, expected)
-		}
-
-		return newErrFixtureMismatch(msg)
-	}
-
-	return nil
-}
-
-// compareTemplate is reading the golden fixture file and compare the stored
-// data with the actual data.
-func (g *goldie) compareTemplate(t *testing.T, name string, data interface{}, actualData []byte) error {
-	expectedDataTmpl, err := ioutil.ReadFile(g.GoldenFileName(t, name))
-
-	if err != nil {
-		if os.IsNotExist(err) {
-			return newErrFixtureNotFound()
-		}
-
-		return fmt.Errorf("Expected %s to be nil", err.Error())
-	}
-
-	missingKey := "error"
-	if g.ignoreTemplateErrors {
-		missingKey = "default"
-	}
-	tmpl, err := template.New("test").Option("missingkey=" + missingKey).Parse(string(expectedDataTmpl))
-	if err != nil {
-		return fmt.Errorf("Expected %s to be nil", err.Error())
-	}
-
-	var expectedData bytes.Buffer
-	err = tmpl.Execute(&expectedData, data)
-	if err != nil {
-		return newErrMissingKey(fmt.Sprintf("Template error: %s", err.Error()))
-	}
-
-	if !bytes.Equal(actualData, expectedData.Bytes()) {
-		msg := "Result did not match the golden fixture. Diff is below:\n\n"
-		actual := string(actualData)
-		expected := expectedData.String()
-
-		if g.diffFn != nil {
-			msg += g.diffFn(actual, expected)
-		} else {
-			msg += Diff(g.diffEngine, actual, expected)
-		}
-
-		return newErrFixtureMismatch(msg)
-	}
-
-	return nil
-}
-
 // ensureDir will create the fixture folder if it does not already exist.
-func (g *goldie) ensureDir(loc string) error {
+func (g *Goldie) ensureDir(loc string) error {
 	s, err := os.Stat(loc)
 
 	switch {
@@ -480,8 +191,7 @@ func (g *goldie) ensureDir(loc string) error {
 }
 
 // GoldenFileName simply returns the file name of the golden file fixture.
-func (g *goldie) GoldenFileName(t *testing.T, name string) string {
-
+func (g *Goldie) GoldenFileName(t *testing.T, name string) string {
 	dir := g.fixtureDir
 
 	if g.useTestNameForDir {
