+++ goldie_test.go
@@ -102,7 +102,8 @@ func TestEnsureDir(t *testing.T) {
 
 			f, err := os.Create(target)
 			require.NoError(t, err)
-			f.Close()
+			err = f.Close()
+			assert.Nil(t, err)
 		}
 
 		err := g.ensureDir(target)
@@ -137,127 +138,6 @@ func TestUpdate(t *testing.T) {
 
 		err = os.RemoveAll(g.fixtureDir)
 		assert.Nil(t, err)
-	}
-}
-
-func TestCompare(t *testing.T) {
-	tests := []struct {
-		name         string
-		actualData   []byte
-		expectedData []byte
-		update       bool
-		err          error
-	}{
-		{
-			name:         "example",
-			actualData:   []byte("abc"),
-			expectedData: []byte("abc"),
-			update:       true,
-			err:          nil,
-		},
-		{
-			name:         "example",
-			actualData:   []byte("abc"),
-			expectedData: []byte("abc"),
-			update:       false,
-			err:          &errFixtureNotFound{},
-		},
-		{
-			name:         "example",
-			actualData:   []byte("bc"),
-			expectedData: []byte("abc"),
-			update:       true,
-			err:          &errFixtureMismatch{},
-		},
-		{
-			name:         "nil",
-			actualData:   nil,
-			expectedData: nil,
-			update:       true,
-			err:          nil,
-		},
-	}
-
-	g := New(t)
-
-	for _, test := range tests {
-		if test.update {
-			err := g.Update(t, test.name, test.expectedData)
-			assert.Nil(t, err)
-		}
-
-		err := g.compare(t, test.name, test.actualData)
-		assert.IsType(t, test.err, err)
-
-		g.GoldenFileName(t, test.name)
-		err = os.RemoveAll(filepath.Dir(g.GoldenFileName(t, test.name)))
-		assert.Nil(t, err)
-	}
-}
-
-func TestCompareTemplate(t *testing.T) {
-	data := struct {
-		Name string
-	}{
-		Name: "example",
-	}
-
-	tests := []struct {
-		name         string
-		actualData   []byte
-		expectedData []byte
-		data         interface{}
-		update       bool
-		err          error
-	}{
-		{
-			name:         "example",
-			actualData:   []byte("abc example"),
-			expectedData: []byte("abc {{ .Name }}"),
-			data:         data,
-			update:       true,
-			err:          nil,
-		},
-		{
-			name:         "example",
-			actualData:   []byte("abc example"),
-			expectedData: []byte("abc {{ .Name }}"),
-			data:         nil,
-			update:       false,
-			err:          &errFixtureNotFound{},
-		},
-		{
-			name:         "example",
-			actualData:   []byte("bc example"),
-			expectedData: []byte("abc {{ .Name }}"),
-			data:         data,
-			update:       true,
-			err:          &errFixtureMismatch{},
-		},
-		{
-			name:         "example",
-			actualData:   []byte("bc example"),
-			expectedData: []byte("abc {{ .Name }}"),
-			data:         nil,
-			update:       true,
-			err:          &errMissingKey{},
-		}}
-
-	g := New(t)
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			if test.update {
-				err := g.Update(t, test.name, test.expectedData)
-				assert.Nil(t, err)
-			}
-
-			err := g.compareTemplate(t, test.name, test.data, test.actualData)
-			assert.IsType(t, test.err, err)
-
-			err = os.RemoveAll(g.fixtureDir)
-			assert.Nil(t, err)
-		})
 	}
 }
 
@@ -339,7 +219,7 @@ func TestCleanFunction(t *testing.T) {
 	}{
 		{fixtureDirWithSub: fixtureSubDirA, filePrefix: "example-a1"},
 		{fixtureDirWithSub: fixtureSubDirA, filePrefix: "example-a2"},
-		{fixtureDirWithSub: fixtureSubDirB, filePrefix: "example-b1",},
+		{fixtureDirWithSub: fixtureSubDirB, filePrefix: "example-b1"},
 		{fixtureDirWithSub: fixtureSubDirB, filePrefix: "example-b2"},
 	}
 
@@ -372,7 +252,7 @@ func TestCleanFunction(t *testing.T) {
 	}{
 		{fixtureDirWithSub: fixtureSubDirA, filePrefix: "example-a3"},
 		{fixtureDirWithSub: fixtureSubDirA, filePrefix: "example-a4"},
-		{fixtureDirWithSub: fixtureSubDirB, filePrefix: "example-b3",},
+		{fixtureDirWithSub: fixtureSubDirB, filePrefix: "example-b3"},
 		{fixtureDirWithSub: fixtureSubDirB, filePrefix: "example-b4"},
 	}
